requirements.txt

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
numpy>=1.24.0,<2.0.0
protobuf<3.20,>=3.9.2  # For descript-audiotools compatibility
scipy>=1.10.0

# Machine learning & embeddings
torch>=2.0.0
transformers>=4.30.0
#openl3>=0.4.0  # Latest available version is 0.4.2

# Vector search
faiss-cpu>=1.7.4

# Data handling
pandas>=2.0.0
pyarrow>=12.0.0

# Visualization & reporting
matplotlib>=3.7.0
seaborn>=0.12.0

# Web UI
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
checksumdir>=1.2.0

# Optional: PDF generation
# weasyprint>=59.0  # or use pandoc/wkhtmltopdf
